+++ pkg/objects/accrual.go
@@ -17,5 +17,5 @@ var AccrualStatusToOrderStatus = map[string]string{
 type Accrual struct {
 	Order   string  `json:"order"`
 	Status  string  `json:"status"`
-	Accrual float64 `json:"accrual,omitempty"`
+	Accrual float32 `json:"accrual,omitempty"`
 }
